add sympy to requirements.txt; move some functions to math_utils.py
